+++ tests/agents/test_estimate_agent.py
@@ -1,0 +1,211 @@
+"""Test :class:`.EstimateAgent`."""
+# Standard Library Imports
+# Third Party Imports
+import pytest
+from numpy import array, diagflat
+
+# RESONAATE Imports
+import resonaate.estimation.initial_orbit_determination
+from resonaate.agents.estimate_agent import EstimateAgent
+from resonaate.agents.sensing_agent import SensingAgent
+from resonaate.data.observation import Observation
+from resonaate.dynamics.terrestrial import Terrestrial
+from resonaate.dynamics.two_body import TwoBody
+from resonaate.estimation import initialOrbitDeterminationFactory
+from resonaate.estimation.maneuver_detection import StandardNis
+from resonaate.estimation.sequential.unscented_kalman_filter import UnscentedKalmanFilter
+from resonaate.physics.time.stardate import JulianDate, ScenarioTime
+from resonaate.scenario.clock import ScenarioClock
+from resonaate.scenario.config.estimation_config import InitialOrbitDeterminationConfig
+from resonaate.sensors.advanced_radar import AdvRadar
+from resonaate.sensors.sensor_base import ObservationTuple
+
+
+@pytest.fixture(name="earth_sensor")
+def getTestEarthSensor():
+    """Create a custom :class:`Agent` object for a sensor."""
+    earth_sensor = AdvRadar(
+        az_mask=array([0.0, 359.99999]),
+        el_mask=array([1.0, 89.0]),
+        r_matrix=diagflat([2.38820057e-11, 3.73156339e-11, 9.00000000e-08, 3.61000000e-10]),
+        diameter=27.0,
+        efficiency=0.9,
+        exemplar=array([1, 40744]),
+        field_of_view={"fov_shape": "conic"},
+        calculate_fov=False,
+        power_tx=120000.0,
+        frequency=10000000000.0,
+        slew_rate=3.0000000000000004,
+        detectable_vismag=25.0,
+        minimum_range=0.0,
+        maximum_range=99000,
+    )
+    return earth_sensor
+
+
+@pytest.fixture(name="sensor_agent")
+def getTestSensorAgent(earth_sensor):
+    """Create a custom :class:`Agent` object for a sensor."""
+    sensor_agent = SensingAgent(
+        300000,
+        "Test_sensor",
+        "GroundFacility",
+        array(
+            [
+                -1.55267475e03,
+                1.47362430e03,
+                5.98812597e03,
+                -1.07453539e-01,
+                -1.14109571e-01,
+                2.19474290e-04,
+            ]
+        ),
+        ScenarioClock(JulianDate(2459304.1666666665), 28800.0, 300),
+        earth_sensor,
+        Terrestrial(
+            JulianDate(2459304.1666666665),
+            array(
+                [
+                    1.83995228e03,
+                    1.11114727e03,
+                    5.98497681e03,
+                    1.16467265e-24,
+                    -6.00704788e-24,
+                    3.01869766e-18,
+                ]
+            ),
+        ),
+        True,
+        10.0,
+        100.0,
+        0.21,
+    )
+    return sensor_agent
+
+
+@pytest.fixture(name="obs_tuples")
+def getObsTuples(sensor_agent):
+    """Create a custom :class:`ObservationTuple` object for a sensor."""
+    radar_observation = Observation.fromSEZVector(
+        julian_date=JulianDate(2459304.267361111),
+        sensor_id=300000,
+        target_id=10001,
+        sensor_type="AdvRadar",
+        azimuth_rad=1.228134787553298,
+        elevation_rad=0.5432822498364404,
+        range_km=1953.3903221962914,
+        range_rate_km_p_sec=-6.147282606743988,
+        sez=[247.45853451777919, 1923.5170160471882, 233.61929237537493],
+        sensor_position=[1.228134787553298, 0.5432822498364407, 0.06300000000101136],
+    )
+
+    return ObservationTuple(radar_observation, sensor_agent, array([2, 3, 1, 1]))
+
+
+@pytest.fixture(name="nominal_filter")
+def getNominalFilter():
+    """Create a :class:`.NominalFilter`."""
+    nominal_filter = UnscentedKalmanFilter(
+        10001,
+        ScenarioTime(0.0),
+        array([6378.0, 2.0, 10.0, 0.0, 7.0, 0.0]),
+        diagflat([1.0, 2.0, 1.0, 1, 1, 1]),
+        TwoBody(),
+        3 * diagflat([1.0, 2.0, 1.0, 1, 1, 1]),
+        StandardNis(0.01),
+        True,
+    )
+    nominal_filter.maneuver_detected = True
+    return nominal_filter
+
+
+@pytest.fixture(name="estimate_agent")
+def getTestEstimateAgent(
+    nominal_filter,
+):
+    """Create a custom :class:.`Agent` object for testing."""
+    estimate_agent = EstimateAgent(
+        10001,
+        "estimate_agent",
+        "Spacecraft",
+        ScenarioClock(JulianDate(2459304.0666666665), 28800.0, 300),
+        array(
+            [
+                4.38022237e03,
+                -5.70920771e01,
+                -5.29317945e03,
+                -5.76471671e00,
+                -1.46542677e00,
+                -4.75396773e00,
+            ]
+        ),
+        diagflat([1.0, 2.0, 1.0, 1, 1, 1]),
+        nominal_filter,
+        None,
+        None,
+        10.0,
+        100.0,
+        0.21,
+    )
+    return estimate_agent
+
+
+def testAttemptInitialOrbitDetermination(estimate_agent, obs_tuples, monkeypatch, caplog):
+    """Test _attemptInitialOrbitDetermination().
+
+    Args:
+        estimate_agent (:class:`.EstimateAgent`): Estimate agent fixture
+        obs_tuples (:class:`.ObservationTuple`): Observation tuple fixture
+        monkeypatch (``:class:`.MonkeyPatch``): patch of function
+        caplog (:class:`.LogCaptureFixture`): logging of function calls
+    """
+    # pylint:disable=protected-access
+    # Test the quick return
+    estimate_agent._attemptInitialOrbitDetermination(obs_tuples)
+    assert estimate_agent.iod_start_time is None
+
+    # Test setting the start time
+    iod_config = InitialOrbitDeterminationConfig()
+    estimate_agent.initial_orbit_determination = initialOrbitDeterminationFactory(
+        iod_config, estimate_agent.simulation_id, estimate_agent.julian_date_start
+    )
+    estimate_agent._attemptInitialOrbitDetermination(obs_tuples)
+    assert estimate_agent.iod_start_time == ScenarioTime(0.0)
+
+    # Test unsuccessful IOD
+    estimate_agent.time = ScenarioTime(300.0)
+
+    def determineNewEstimateStateBad(
+        self, obs_tuples, detection_time, current_time
+    ):  # pylint:disable=unused-argument
+        return array([1, 2, 3, 4, 5, 7]), False
+
+    with monkeypatch.context() as m_patch:
+        m_patch.setattr(
+            resonaate.estimation.initial_orbit_determination.LambertIOD,
+            "determineNewEstimateState",
+            determineNewEstimateStateBad,
+        )
+        estimate_agent._attemptInitialOrbitDetermination(obs_tuples)
+        assert (
+            caplog.record_tuples[-1][-1]
+            == f"IOD unsuccessful for RSO {estimate_agent.simulation_id}"
+        )
+
+    # Test successful IOD
+    def determineNewEstimateStateGood(
+        self, obs_tuples, detection_time, current_time
+    ):  # pylint:disable=unused-argument
+        return array([1, 2, 3, 4, 5, 6]), True
+
+    with monkeypatch.context() as m_patch:
+        m_patch.setattr(
+            resonaate.estimation.initial_orbit_determination.LambertIOD,
+            "determineNewEstimateState",
+            determineNewEstimateStateGood,
+        )
+        estimate_agent._attemptInitialOrbitDetermination(obs_tuples)
+        assert (
+            caplog.record_tuples[-1][-1]
+            == f"IOD successful for RSO {estimate_agent.simulation_id}"
+        )